+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - 🎉 Added `datatype.octal`
 - 🎉 Added `datatype.numeric`
+- The prefix argument was added to the `datatype.alphaNumeric` module
 
 ## 🪛 Fix